refactor resource for cloudflare
