+++ sdk/js/src/auth/session.ts
@@ -1,5 +1,5 @@
 import { SignJWT, importPKCS8, importSPKI, jwtVerify } from "jose";
-import { Resource } from "../resource.js";
+import { Resource } from "../resource/index.js";
 import process from "node:process";
 
 export type SessionBuilder = ReturnType<typeof createSessionBuilder>;
