+++ sdk/js/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./realtime/index.js";
-export * from "./resource.js";
+export * from "./resource/index.js";
 export * from "./vector/index.js";
 
 import { format } from "util";
