+++ sdk/js/src/resource/cloudflare.ts
@@ -1,0 +1,7 @@
+import { env } from "cloudflare:workers";
+
+import { fromCloudflareEnv, Resource, wrapCloudflareHandler } from "./shared.js";
+
+fromCloudflareEnv(env);
+
+export { Resource, fromCloudflareEnv, wrapCloudflareHandler };
