+++ sdk/js/src/resource/workers.d.ts
@@ -1,0 +1,3 @@
+declare module "cloudflare:workers" {
+  export const env: Record<string, any>;
+}
